+++ DataManipulator/CMakeLists.txt
@@ -2,7 +2,26 @@
 
 project ("DataManipulator")
 
-add_executable (DataManipulator "DataManipulator.cpp" "DataManipulator.h" "DataFormats.h" "DataFormats.cpp" "WARNING.h" "WARNING.cpp" "DataReformer.h" "DataReformer.cpp")
+set(SOURCE_FILES
+        DataManipulator.h
+        DataManipulator.cpp
+        MainOCDFMenu.h
+        MainOCDFMenu.cpp
+        TDFCreator.h
+        TDFCreator.cpp
+        ../Common/DataFormats.h
+        ../Common/DataFormats.cpp
+        ../Common/WARNING.h
+        ../Common/WARNING.cpp
+        ../Common/DataReformer.h
+        ../Common/DataReformer.cpp
+        ../Common/DataReader.h
+        ../Common/DataReader.cpp
+        ../Common/DataDumper.h
+        ../Common/DataDumper.cpp
+        )
+
+add_executable (DataManipulator ${SOURCE_FILES})
 
 if (CMAKE_VERSION VERSION_GREATER 3.12)
   set_property(TARGET DataManipulator PROPERTY CXX_STANDARD 20)
